// frontend/src/app/api/model/plan.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanItem } from './planItem';
import { PlanStatus } from './planStatus';


export interface Plan { 
    id: number;
    title: string;
    items: Array<PlanItem>;
    estimatedDurationMinutes: number;
    startDateTime: string;
    endDateTime: string;
    priorityOrder: number;
    /**
     * Computed on read (FR-08.4, Q2)
     */
    status: PlanStatus;
    createdAt: string;
    doneItems: number;
    totalItems: number;
    progressPercent: number;
    /**
     * Seconds until endDateTime while start <= now < end, otherwise null (BR-12)
     */
    restSeconds?: number | null;
}
export namespace Plan {
}