// frontend/src/app/api/model/planSourceType.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PlanSourceType = {
    Task: 'TASK',
    Habit: 'HABIT',
    LearningResource: 'LEARNING_RESOURCE',
} as const;
export type PlanSourceType = typeof PlanSourceType[keyof typeof PlanSourceType];